removed collections in entities

+++ src/main/java/library/service/impl/BookServiceImpl.java
@@ -1,32 +1,58 @@
 package library.service.impl;
 
+import library.dto.AuthorDto;
+import library.dto.AuthorInBookDto;
+import library.dto.BookDto;
 import library.exception.AlreadyExistException;
 import library.exception.BadRequestParametrException;
 import library.exception.NotAllAuthorsDataException;
 import library.exception.NotFoundException;
-import library.model.Author;
 import library.model.Book;
+import library.repository.AuthorRepository;
+import library.repository.BookAuthorRepository;
 import library.repository.BookRepository;
+import library.repository.LoanRepository;
 import library.service.AuthorService;
+import library.service.BookAuthorService;
 import library.service.BookService;
-import org.springframework.beans.factory.annotation.Autowired;
+import library.transformer.AuthorTransformer;
+import library.transformer.BookTransformer;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.ArrayList;
-import java.util.Collection;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 @Service
 public class BookServiceImpl implements BookService {
     private final BookRepository bookRepository;
+    private final AuthorRepository authorRepository;
+    private final BookAuthorRepository bookAuthorRepository;
+    private final AuthorTransformer authorTransformer;
     private final AuthorService authorService;
+    private final BookTransformer bookTransformer;
+    private final BookAuthorService bookAuthorService;
+    private final LoanRepository loanRepository;
 
-    @Autowired
-    public BookServiceImpl(BookRepository bookRepository, AuthorService authorService) {
+    public BookServiceImpl(BookRepository bookRepository,
+                           AuthorRepository authorRepository,
+                           BookAuthorRepository bookAuthorRepository,
+                           AuthorTransformer authorTransformer,
+                           AuthorService authorService,
+                           BookTransformer bookTransformer,
+                           BookAuthorService bookAuthorService,
+                           LoanRepository loanRepository) {
         this.bookRepository = bookRepository;
+        this.authorRepository = authorRepository;
+        this.bookAuthorRepository = bookAuthorRepository;
+        this.authorTransformer = authorTransformer;
         this.authorService = authorService;
+        this.bookTransformer = bookTransformer;
+        this.bookAuthorService = bookAuthorService;
+        this.loanRepository = loanRepository;
     }
+
 
     @Transactional
     @Override
@@ -36,35 +62,36 @@ public class BookServiceImpl implements BookService {
 
     @Override
     public Book findByName(String bookName) {
-        return bookRepository.findByName(bookName).orElseThrow(()->new NotFoundException("Book", "name", bookName));
+        return bookRepository.findByName(bookName).orElseThrow(() -> new NotFoundException("Book", "name", bookName));
     }
 
     @Transactional
     @Override
-    public Book add(Book book) {
-        Collection<Author> authors = book.getAuthors();
-        String name = book.getName();
+    public BookDto add(BookDto bookDto) {
+        String name = bookDto.getName();
 
         if (bookRepository.existsByName(name)) {
             throw new AlreadyExistException("Book", "name", name);
         }
+        Set<Long> authorsToSaveId = new HashSet<>();
 
-        List<Author> authorsToSave = new ArrayList<>();
-        for (Author author : authors) {
-            Author existingAuthor;
-            if (authorService.existsByName(author.getName())) {
-                existingAuthor = authorService.findByName(author.getName());
-                existingAuthor.getBooks().add(book);
-                authorsToSave.add(existingAuthor);
+        Set<AuthorInBookDto> authorsToSave = bookDto.getAuthors();
+        for (AuthorInBookDto author : authorsToSave) {
+            if (authorRepository.existsByName(author.getName())) {
+                authorsToSaveId.add(authorRepository.findByName(author.getName()).get().getId());
             } else if (author.getName() == null || author.getBirthday() == null) {
                 throw new NotAllAuthorsDataException();
             } else {
-                authorsToSave.add(author);
+                AuthorDto authorDto = new AuthorDto();
+                authorDto.setName(author.getName());
+                authorDto.setBirthday(author.getBirthday());
+                authorDto = authorService.add(authorDto);
+                authorsToSaveId.add(authorDto.getId());
             }
         }
-        book.setInStockQuantity(book.getTotalQuantity());
-        book.setAuthors(authorsToSave);
-        return bookRepository.save(book);
+        Book book = bookRepository.save(bookTransformer.toBook(bookDto));
+        bookAuthorService.saveAuthorsInBook(authorsToSaveId, book.getId());
+        return bookTransformer.toBookDto(book);
     }
 
     @Override
@@ -75,10 +102,12 @@ public class BookServiceImpl implements BookService {
     @Transactional
     @Override
     public void deleteById(Long id) {
-        Book book = findById(id);
-        book.setAuthors(null);
-        bookRepository.save(book);
+        findById(id);
+        if (loanRepository.existsByBookId(id)) {
+            throw new BadRequestParametrException("Not allowed delete a book that is loaned.");
+        }
         bookRepository.deleteById(id);
+        bookAuthorRepository.deleteAllByBookId(id);
     }
 
     @Override
@@ -89,24 +118,41 @@ public class BookServiceImpl implements BookService {
     @Override
     public Integer getQuantityInStockByName(String name) {
         return bookRepository.findByName(name)
-                .orElseThrow(()->new NotFoundException("Book", "name", name))
+                .orElseThrow(() -> new NotFoundException("Book", "name", name))
                 .getInStockQuantity();
     }
 
     @Transactional
     @Override
-    public Book update(Book book) {
-        if (existByName(book.getName())) {
-            throw new AlreadyExistException("Book", "name", book.getName());
-        }
-        Book updatableBook = findById(book.getId());
+    public BookDto update(BookDto bookDto) {
+        String name = bookDto.getName();
 
-        updatableBook.setName(book.getName());
-        updatableBook.setPublisher(book.getPublisher());
-        updatableBook.setTotalQuantity(book.getTotalQuantity());
-        updatableBook.setInStockQuantity(book.getInStockQuantity());
-        updatableBook.setAuthors(book.getAuthors());
-        return bookRepository.save(updatableBook);
+        if (existByName(name)&&
+                (bookRepository.findByName(name).get().getId()!=bookDto.getId())) {
+            throw new AlreadyExistException("Book", "name", bookDto.getName());
+        }
+        Long bookId = bookDto.getId();
+        Book updatableBook = findById(bookId);
+
+        updatableBook.setName(bookDto.getName());
+        updatableBook.setPublisher(bookDto.getPublisher());
+        updatableBook.setTotalQuantity(bookDto.getTotalQuantity());
+        updatableBook.setInStockQuantity(bookDto.getInStockQuantity());
+
+        bookAuthorRepository.deleteAllByBookId(bookId);
+
+        Set<AuthorInBookDto> authorInBookDto = bookDto.getAuthors();
+        for (AuthorInBookDto author : authorInBookDto) {
+            if (authorRepository.existsByName(author.getName())) {
+                Long authorId = authorRepository.findByName(author.getName()).get().getId();
+                bookAuthorService.saveAuthorInBook(authorId, bookId);
+            } else {
+                AuthorDto authorDto = authorTransformer.toAuthorDto(author);
+                authorDto = authorService.add(authorDto);
+                bookAuthorService.saveAuthorInBook(authorDto.getId(), bookId);
+            }
+        }
+        return bookTransformer.toBookDto(bookRepository.save(updatableBook));
     }
 
     @Transactional
@@ -118,7 +164,7 @@ public class BookServiceImpl implements BookService {
         if (getQuantityInStockByName(bookName) < quantity) {
             throw new BadRequestParametrException("Less books in the library than want to take.");
         }
-        Book takedBook =bookRepository.findByName(bookName).orElseThrow(()->new NotFoundException("Book", "name", bookName));
+        Book takedBook = bookRepository.findByName(bookName).orElseThrow(() -> new NotFoundException("Book", "name", bookName));
         Integer newQuantity = takedBook.getInStockQuantity() - quantity;
         takedBook.setInStockQuantity(newQuantity);
         bookRepository.save(takedBook);
@@ -127,15 +173,9 @@ public class BookServiceImpl implements BookService {
     @Transactional
     @Override
     public void returnBook(String bookName, Integer quantity) {
-        Book takedBook = bookRepository.findByName(bookName).orElseThrow(()->new NotFoundException("Book", "name", bookName));
+        Book takedBook = bookRepository.findByName(bookName).orElseThrow(() -> new NotFoundException("Book", "name", bookName));
         Integer newQuantity = takedBook.getInStockQuantity() + quantity;
         takedBook.setInStockQuantity(newQuantity);
         bookRepository.save(takedBook);
-    }
-
-    @Transactional
-    @Override
-    public void removeAuthor(Book book, Author deletedAuthor) {
-        book.getAuthors().remove(deletedAuthor);
     }
 }
